+++ src/constants.ts
@@ -1,0 +1,18 @@
+export enum OUTPUT_FORMAT {
+  SCSS = 'SCSS',
+  LESS = 'LESS',
+  CSS = 'CSS'
+  // TODO enable json output
+}
+
+export enum NAME_FORMAT {
+  CAMEL_HYPHEN = 'camel-hyphen',
+  LOWCASE_HYPHEN = 'lowcase-hyphen'
+}
+
+export enum COMMAND_TYPE {
+  COPY = 'COPY',
+  GENERATE_CODE = 'GENERATE_CODE',
+  DOWNLOAD = 'DOWNLOAD',
+  CLEAN = 'CLEAN'
+}