+++ src/plugin.ts
@@ -1,28 +1,28 @@
-import { formatVariable, getMixinPrefix, getColorValue } from "./utils";
-import { OutputFormat, CommandType, IMessageFormat } from "./interfaces";
+import { Utilities } from './utilities';
+import { IMessageFormat } from './interfaces';
+import { OUTPUT_FORMAT, COMMAND_TYPE } from './constants';
 
 let count = 0;
 let colorStyles = {};
 let textStyles = {};
-let format: OutputFormat = "scss";
-let currentGeneratedCode: string = "";
+let format: OUTPUT_FORMAT = OUTPUT_FORMAT.SCSS;
+let currentGeneratedCode: string = '';
 
 // Display UI
 figma.showUI(__html__, {
-  height: 300,
-  width: 300
+  height: 616,
+  width: 500
 });
 
 // Define listener to message from the ui and choose what to do based on the command type
 figma.ui.onmessage = message => {
   switch (message.command) {
-    case CommandType.DOWNLOAD:
+    case COMMAND_TYPE.DOWNLOAD:
       download(message);
       break;
-    case CommandType.GENERATE_CODE:
+    case COMMAND_TYPE.GENERATE_CODE:
       generateCode(message);
       break;
-
     default:
       break;
   }
@@ -37,8 +37,8 @@ function download(message: IMessageFormat) {
 // Iterate over the figma tree and fetch the styles.
 // Current support is text styles and color styles
 function generateCode(message: IMessageFormat) {
-  figma.ui.postMessage({ command: CommandType.CLEAN });
-  let generatedCode = "";
+  figma.ui.postMessage({ command: COMMAND_TYPE.CLEAN });
+  let generatedCode = '';
 
   if (message.format) {
     format = message.format;
@@ -46,27 +46,30 @@ function generateCode(message: IMessageFormat) {
 
   // Release the ui thread to paint and exec traverse.
   setTimeout(() => {
-    if (typeof figma.getLocalPaintStyles === "function") {
+    if (typeof figma.getLocalPaintStyles === 'function') {
       getLocalStyles();
     } else {
       traverse(figma.root);
     }
-    console.log("count", count);
-    if (format === "css") {
+    console.log('count', count);
+    if (format === OUTPUT_FORMAT.CSS) {
       generatedCode += `:root {\n`;
     }
 
     for (const key in colorStyles) {
       const val = colorStyles[key];
-      const preprocessorVariable = `${formatVariable(key, format)}:${val};\n`;
+      const preprocessorVariable = `${Utilities.getVariablePrefix(format)}color-${Utilities.formatVariable(
+        key,
+        format
+      )}: ${val};\n`;
       generatedCode += preprocessorVariable;
     }
 
-    if (format !== "css") {
+    if (format !== OUTPUT_FORMAT.CSS) {
       for (const key in textStyles) {
         const element = textStyles[key];
-        const mixinName = formatVariable(key, format).replace(/^\$|\@/g, "");
-        let value: string = `${getMixinPrefix(format, mixinName)} {\n`;
+        const mixinName = Utilities.formatVariable(key, format).replace(/^\$|\@/g, '');
+        let value: string = `${Utilities.getMixinPrefix(format, mixinName)} {\n`;
         for (const cssRule in element) {
           const cssValue = element[cssRule];
           value += `\t${cssRule}:${cssValue};\n`;
@@ -88,12 +91,12 @@ function getLocalStyles() {
   const paintStyles = figma.getLocalPaintStyles();
   paintStyles.forEach((style: PaintStyle) => {
     // Prepare style
-    if (!style.paints || !style.paints.length || !style.paints[0]["color"]) {
+    if (!style.paints || !style.paints.length || !style.paints[0]['color']) {
       return;
     }
-    const color = style.paints[0]["color"];
+    const color = style.paints[0]['color'];
     const opacity = style.paints[0].opacity;
-    const val = getColorValue(color, opacity);
+    const val = Utilities.getColorValue(color, opacity !== undefined ? opacity : 1);
 
     // Count styles
     colorStyles[style.name] = val;
@@ -104,18 +107,18 @@ function getLocalStyles() {
   tStyles.forEach((style: TextStyle) => {
     let textValues = {};
     if (style.fontSize) {
-      textValues["font-size"] = `${style.fontSize}px`;
+      textValues['font-size'] = `${style.fontSize}px`;
     }
     if (style.fontName && style.fontName.style) {
-      textValues["font-family"] = `"${style.fontName.family}"`;
+      textValues['font-family'] = `"${style.fontName.family}"`;
       const fontStyle = style.fontName.style.toLowerCase();
-      const fontStyleValue = fontStyle === "regular" ? "normal" : fontStyle;
-      textValues["font-style"] = fontStyleValue;
+      const fontStyleValue = fontStyle === 'regular' ? 'normal' : fontStyle;
+      textValues['font-weight'] = fontStyleValue;
     }
 
     // TODO by UNIT
-    if (style.lineHeight && style.lineHeight["value"]) {
-      textValues["line-height"] = style.lineHeight["value"] + "px";
+    if (style.lineHeight && style.lineHeight['value']) {
+      textValues['line-height'] = style.lineHeight['value'] + 'px';
     }
     textStyles[style.name] = textValues;
     count++;
@@ -127,11 +130,11 @@ function traverse(node: BaseNode) {
     return;
   }
 
-  if ("children" in node) {
-    if (node.type !== "INSTANCE") {
+  if ('children' in node) {
+    if (node.type !== 'INSTANCE') {
       for (const child of node.children) {
         traverse(child);
-        if (child.type === "FRAME") {
+        if (child.type === 'FRAME') {
           const styleId: any = child.backgroundStyleId;
           const style = figma.getStyleById(styleId);
 
@@ -139,23 +142,23 @@ function traverse(node: BaseNode) {
             continue;
           }
 
-          console.log("FRAME", style);
+          console.log('FRAME', style);
           const key = style.name;
           if (colorStyles[key]) {
             continue;
           }
 
           // Prepare style
-          const color = style["paints"][0].color;
-          const opacity = style["paints"][0].opacity;
-          const val = getColorValue(color, opacity);
+          const color = style['paints'][0].color;
+          const opacity = style['paints'][0].opacity;
+          const val = Utilities.getColorValue(color, opacity);
 
           // Count styles
           colorStyles[key] = val;
           count++;
         }
 
-        if (child.type === "RECTANGLE" && child.fillStyleId) {
+        if (child.type === 'RECTANGLE' && child.fillStyleId) {
           const styleId: any = child.fillStyleId;
           const style = figma.getStyleById(styleId);
 
@@ -169,19 +172,19 @@ function traverse(node: BaseNode) {
           }
 
           // Prepare style
-          const color = style["paints"][0].color;
-          const opacity = style["paints"][0].opacity;
-          const val = getColorValue(color, opacity);
+          const color = style['paints'][0].color;
+          const opacity = style['paints'][0].opacity;
+          const val = Utilities.getColorValue(color, opacity);
 
           // Count styles
           colorStyles[key] = val;
           count++;
         }
 
-        if (child.type === "TEXT") {
+        if (child.type === 'TEXT') {
           console.log(child);
           const styleId: any = child.textStyleId;
-          if (!styleId || typeof styleId !== "string") {
+          if (!styleId || typeof styleId !== 'string') {
             continue;
           }
 
@@ -199,13 +202,13 @@ function traverse(node: BaseNode) {
           }
 
           let textValues = {};
-          textValues["font-size"] = `${style.fontSize}px`;
-          textValues["font-family"] = `"${style.fontName.family}"`;
+          textValues['font-size'] = `${style.fontSize}px`;
+          textValues['font-family'] = `"${style.fontName.family}"`;
           const fontStyle = style.fontName.style.toLowerCase();
-          const fontStyleValue = fontStyle === "regular" ? "normal" : fontStyle;
-          textValues["font-style"] = fontStyleValue;
+          const fontStyleValue = fontStyle === 'regular' ? 'normal' : fontStyle;
+          textValues['font-style'] = fontStyleValue;
           // TODO by UNIT
-          textValues["line-height"] = style.lineHeight["value"] + "px";
+          textValues['line-height'] = style.lineHeight['value'] + 'px';
           textStyles[key] = textValues;
           count++;
         }