+++ src/ui.ts
@@ -1,7 +1,0 @@
-import "../node_modules/prismjs/themes/prism.css";
-import "../node_modules/file-saver/dist/FileSaver.js";
-import "../node_modules/prismjs";
-import "../node_modules/prismjs/components/prism-scss";
-import "../node_modules/prismjs/components/prism-less";
-import "../node_modules/prismjs/components/prism-stylus";
-import "./ui.scss";