+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.1.1](https://github.com/noveo-io/figma-sass-less-plugin/compare/v1.1.0...v1.1.1) (2019-11-06)
+
+
+### Bug Fixes
+
+* **constants:** remove a comment ([0ce3338](https://github.com/noveo-io/figma-sass-less-plugin/commit/0ce333887956bf6b13d0fd90c4c34966a25dbfdc))
+
+## [1.1.0](https://github.com/noveo-io/figma-sass-less-plugin/compare/v1.0.3...v1.1.0) (2019-11-06)
+
+
+### Features
+
+* **plugin:** add kebab-case formatting ([92b5677](https://github.com/noveo-io/figma-sass-less-plugin/commit/92b567717347447deaeef08ac99b069d38a091ba))
+* **plugin:** add names formatting option ([3d6404e](https://github.com/noveo-io/figma-sass-less-plugin/commit/3d6404ed08b98210cf7c9259742c5e8217c006e1))
+
+
+### Bug Fixes
+
+* **plugin:** font weight instead of font style ([55eed8b](https://github.com/noveo-io/figma-sass-less-plugin/commit/55eed8be758cd48945b4896fc25ce3529fd17fe3))
+
 ### [1.0.3](https://github.com/noveo-io/figma-sass-less-plugin/compare/v1.0.1...v1.0.3) (2019-10-18)
 
 ### [1.0.2](https://github.com/noveo-io/figma-sass-less-plugin/compare/v1.0.1...v1.0.2) (2019-10-18)